Started rewriting Sonos provider.
