+++ music_assistant/providers/sonos/provider.py
@@ -207,36 +207,36 @@ class SonosPlayerProvider(PlayerProvider):
                 await airplay_prov.cmd_stop(airplay_player.player_id)
                 airplay_player.active_source = None
 
-    async def cmd_stop(self, player_id: str) -> None:
-        """Send STOP command to given player."""
-        if sonos_player := self.sonos_players[player_id]:
-            await sonos_player.cmd_stop()
+    # async def cmd_stop(self, player_id: str) -> None:
+    #     """Send STOP command to given player."""
+    #     if sonos_player := self.sonos_players[player_id]:
+    #         await sonos_player.cmd_stop()
 
-    async def cmd_play(self, player_id: str) -> None:
-        """Send PLAY command to given player."""
-        if sonos_player := self.sonos_players[player_id]:
-            await sonos_player.cmd_play()
+    # async def cmd_play(self, player_id: str) -> None:
+    #     """Send PLAY command to given player."""
+    #     if sonos_player := self.sonos_players[player_id]:
+    #         await sonos_player.cmd_play()
 
-    async def cmd_pause(self, player_id: str) -> None:
-        """Send PAUSE command to given player."""
-        if sonos_player := self.sonos_players[player_id]:
-            await sonos_player.cmd_pause()
+    # async def cmd_pause(self, player_id: str) -> None:
+    #     """Send PAUSE command to given player."""
+    #     if sonos_player := self.sonos_players[player_id]:
+    #         await sonos_player.cmd_pause()
 
-    async def cmd_seek(self, player_id: str, position: int) -> None:
-        """Handle SEEK command for given player.
+    # async def cmd_seek(self, player_id: str, position: int) -> None:
+    #     """Handle SEEK command for given player.
 
-        - player_id: player_id of the player to handle the command.
-        - position: position in seconds to seek to in the current playing item.
-        """
-        if sonos_player := self.sonos_players[player_id]:
-            await sonos_player.cmd_seek(position)
+    #     - player_id: player_id of the player to handle the command.
+    #     - position: position in seconds to seek to in the current playing item.
+    #     """
+    #     if sonos_player := self.sonos_players[player_id]:
+    #         await sonos_player.cmd_seek(position)
 
-    async def cmd_volume_set(self, player_id: str, volume_level: int) -> None:
-        """Send VOLUME_SET command to given player."""
-        if sonos_player := self.sonos_players[player_id]:
-            await sonos_player.cmd_volume_set(volume_level)
+    # async def cmd_volume_set(self, player_id: str, volume_level: int) -> None:
+    #     """Send VOLUME_SET command to given player."""
+    #     if sonos_player := self.sonos_players[player_id]:
+    #         await sonos_player.cmd_volume_set(volume_level)
 
-    async def cmd_volume_mute(self, player_id: str, muted: bool) -> None:
+    # async def cmd_volume_mute(self, player_id: str, muted: bool) -> None:
         """Send VOLUME MUTE command to given player."""
         if sonos_player := self.sonos_players[player_id]:
             await sonos_player.cmd_volume_mute(muted)
@@ -337,15 +337,15 @@ class SonosPlayerProvider(PlayerProvider):
             media.uri, {"name": media.title, "type": "track"}
         )
 
-    async def cmd_next(self, player_id: str) -> None:
-        """Handle NEXT TRACK command for given player."""
-        if sonos_player := self.sonos_players[player_id]:
-            await sonos_player.client.player.group.skip_to_next_track()
+    # async def cmd_next(self, player_id: str) -> None:
+    #     """Handle NEXT TRACK command for given player."""
+    #     if sonos_player := self.sonos_players[player_id]:
+    #         await sonos_player.client.player.group.skip_to_next_track()
 
-    async def cmd_previous(self, player_id: str) -> None:
-        """Handle PREVIOUS TRACK command for given player."""
-        if sonos_player := self.sonos_players[player_id]:
-            await sonos_player.client.player.group.skip_to_previous_track()
+    # async def cmd_previous(self, player_id: str) -> None:
+    #     """Handle PREVIOUS TRACK command for given player."""
+    #     if sonos_player := self.sonos_players[player_id]:
+    #         await sonos_player.client.player.group.skip_to_previous_track()
 
     async def enqueue_next_media(self, player_id: str, media: PlayerMedia) -> None:
         """Handle enqueuing of the next queue item on the player."""
@@ -373,10 +373,10 @@ class SonosPlayerProvider(PlayerProvider):
         duration = media_info.duration or 10
         await asyncio.sleep(duration)
 
-    async def select_source(self, player_id: str, source: str) -> None:
-        """Handle SELECT SOURCE command on given player."""
-        if sonos_player := self.sonos_players[player_id]:
-            await sonos_player.select_source(source)
+    # async def select_source(self, player_id: str, source: str) -> None:
+    #     """Handle SELECT SOURCE command on given player."""
+    #     if sonos_player := self.sonos_players[player_id]:
+    #         await sonos_player.select_source(source)
 
     async def _setup_player(self, player_id: str, name: str, info: AsyncServiceInfo) -> None:
         """Handle setup of a new player that is discovered using mdns."""
