+++ requirements_all.txt
@@ -30,7 +30,7 @@ ifaddr==0.2.0
 liblistenbrainz==0.5.6
 mashumaro==3.16
 music-assistant-frontend==2.15.2
-music-assistant-models==1.1.47
+music-assistant-models==1.1.48
 mutagen==1.47.0
 orjson==3.10.18
 pillow==11.3.0
